pyproject.toml
[project]
name = "cellcraft"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "numpy>=1.26.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["cellcraft*"]

[tool.ruff]
line-length = 100
